backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
sqlalchemy==2.0.36
pydantic-settings==2.7.0
apscheduler==3.10.4
httpx==0.28.1
playwright==1.49.1

